requirements.txt
albumentations==1.3.0
huggingface-hub==0.10.1
matplotlib==3.5.3
numpy==1.21.6
opencv-python==4.6.0.66
packaging==21.3
pandas==1.3.5
scikit-image==0.19.3
scikit-learn==1.0.2
scipy==1.7.3
urllib3==1.26.12
wandb==0.13.4
torchio
git+https://github.com/shijianjian/EfficientNet-PyTorch-3D